finalize changes
