+++ Components/Main/LineBlock.js
@@ -15,18 +15,18 @@ const LineBlock = (props) => {
   return (
     <View style={style.View}>
       <View stlye={style.column}>
-        <Text style={[style.Gurmukhi, { fontSize: ViewerCtx.gurmukhiSize }]}>{Gurmukhi}</Text>
+        <Text style={[style.Gurmukhi, style.text, { fontSize: ViewerCtx.gurmukhiSize }]}>{Gurmukhi}</Text>
         {
           displayEngTransl &&
-          <Text style={[style.Translation, { fontSize: ViewerCtx.translationSize }]}>{English}</Text>
+          <Text style={[style.Translation, style.text, { fontSize: ViewerCtx.translationSize }]}>{English}</Text>
         }
         {
           displayPunTansl &&
-          <Text style={[style.PunjabiTranslation, { fontSize: ViewerCtx.translationSize }]}>{Punjabi}</Text>
+          <Text style={[style.PunjabiTranslation, style.text, { fontSize: ViewerCtx.translationSize }]}>{Punjabi}</Text>
         }
         {
           displayTranslit &&
-          <Text style={[style.Translation, { fontSize: ViewerCtx.translationSize }]}>{Transliteration.English}</Text>
+          <Text style={[style.Translation, style.text, { fontSize: ViewerCtx.translationSize }]}>{Transliteration.English}</Text>
         }
       </View>
     </View>
@@ -37,25 +37,28 @@ const style = StyleSheet.create({
   View: {
     flexDirection: 'row',
   },
+  text: {
+    paddingLeft: 5,
+  },
   column: {
     flexDirection: 'column',
   },
   Gurmukhi: {
-    marginVertical: 10,
+    marginVertical: 8,
     fontWeight: "400",
     fontFamily: "AnmolLipiTrue",
   },
   Translation: {
-    marginVertical: 5,
+    marginVertical: 3,
     fontWeight: "200",
   },
   PunjabiTranslation: {
-    marginVertical: 5,
+    marginVertical: 3,
     fontWeight: "200",
     fontFamily: "AnmolLipiTrue",
   },
   Translit: {
-    marginVertical: 5,
+    marginVertical: 3,
     fontWeight: "200",
   }
 })
